+++ alfred-telemetry-platform/src/main/java/eu/xenit/alfred/telemetry/binder/DataSourceMetrics.java
@@ -1,16 +1,15 @@
 package eu.xenit.alfred.telemetry.binder;
 
+import eu.xenit.alfred.telemetry.binder.dbcp.TelemetryBasicDataSource;
 import io.micrometer.core.instrument.Gauge;
 import io.micrometer.core.instrument.MeterRegistry;
-import java.util.function.ToDoubleFunction;
-import javax.annotation.Nonnull;
-import javax.sql.DataSource;
-import org.apache.commons.dbcp.BasicDataSource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class DataSourceMetrics implements NamedMeterBinder {
+import java.util.function.ToDoubleFunction;
+import javax.annotation.Nonnull;
 
+public class DataSourceMetrics implements NamedMeterBinder {
     private static final Logger LOGGER = LoggerFactory.getLogger(DataSourceMetrics.class);
 
     @Override
@@ -18,61 +17,44 @@ public class DataSourceMetrics implements NamedMeterBinder {
         return "jdbc";
     }
 
-    private final BasicDataSource dataSource;
+    private final TelemetryBasicDataSource basicDataSource;
 
-    public DataSourceMetrics(DataSource dataSource) {
-        if (dataSource instanceof BasicDataSource) {
-            this.dataSource = (BasicDataSource) dataSource;
-        } else {
-            LOGGER.warn("Monitoring not supported for dataSource of type '{}', expected dataSource of type '{}'",
-                    dataSource.getClass().getCanonicalName(),
-                    BasicDataSource.class.getCanonicalName());
-            this.dataSource = null;
-        }
+    public DataSourceMetrics(TelemetryBasicDataSource versionSpecificBasicDataSource) {
+        this.basicDataSource = versionSpecificBasicDataSource;
     }
 
     @Override
     public void bindTo(@Nonnull MeterRegistry meterRegistry) {
-        if (dataSource == null) {
+        if (basicDataSource == null) {
+            LOGGER.debug("basicDataSource is null, aborting!");
             return;
         }
 
-        dataSourceGaugeBuilder("jdbc.connections.count", BasicDataSource::getNumActive)
+        dataSourceGaugeBuilder("jdbc.connections.count", TelemetryBasicDataSource::getNumActive)
                 .tags("status", "active")
                 .description("The current number of active connections that have been allocated from this data source.")
                 .register(meterRegistry);
-        dataSourceGaugeBuilder("jdbc.connections.count", BasicDataSource::getNumIdle)
+        dataSourceGaugeBuilder("jdbc.connections.count", TelemetryBasicDataSource::getNumIdle)
                 .tags("status", "idle")
                 .description(
                         "The current number of idle connections that are waiting to be allocated from this data source.")
                 .register(meterRegistry);
 
-        dataSourceGaugeBuilder("jdbc.connections.max", BasicDataSource::getMaxActive)
+        dataSourceGaugeBuilder("jdbc.connections.max", TelemetryBasicDataSource::getMaxActive)
                 .description("The maximum number of active connections that can be allocated at the same time.")
                 .register(meterRegistry);
 
-        dataSourceGaugeBuilder("jdbc.connections.min", BasicDataSource::getMinIdle)
+        dataSourceGaugeBuilder("jdbc.connections.min", TelemetryBasicDataSource::getMinIdle)
                 .description("The minimum number of idle connections in the pool.")
                 .register(meterRegistry);
 
-        dataSourceGaugeBuilder("jdbc.connections.usage", this::getUsage)
+        dataSourceGaugeBuilder("jdbc.connections.usage", TelemetryBasicDataSource::getUsage)
                 .register(meterRegistry);
     }
 
-    private Gauge.Builder dataSourceGaugeBuilder(String metricName, ToDoubleFunction<BasicDataSource> function) {
-        return Gauge.builder(metricName, dataSource, function)
+    private Gauge.Builder<TelemetryBasicDataSource> dataSourceGaugeBuilder(String metricName,
+                                                                           ToDoubleFunction<TelemetryBasicDataSource> function) {
+        return Gauge.builder(metricName, basicDataSource, function)
                 .tags("name", "dbcp");
-    }
-
-    private Float getUsage(BasicDataSource dataSource) {
-        int maxSize = dataSource.getMaxActive();
-        int currentSize = dataSource.getNumActive();
-        if (maxSize < 0) {
-            return -1F;
-        }
-        if (currentSize == 0) {
-            return 0F;
-        }
-        return (float) currentSize / (float) maxSize;
     }
 }