+++ alfred-telemetry-platform/src/main/java/eu/xenit/alfred/telemetry/binder/cache/CacheMetricsFactory.java
@@ -30,7 +30,7 @@ public class CacheMetricsFactory extends AbstractFactoryBean<MeterBinder> implem
     @Nonnull
     protected MeterBinder createInstance() {
         return isEnterpriseEdition()
-                ? new EnterpriseCacheMetrics(ctx, descriptorService)
+                ? new EnterpriseCacheMetrics(ctx)
                 : new CommunityCacheMetrics(ctx);
     }