+++ alfred-telemetry-platform/src/main/java/eu/xenit/alfred/telemetry/binder/cache/EnterpriseCacheMetrics.java
@@ -8,34 +8,36 @@ import io.micrometer.core.instrument.MeterRegistry;
 import io.micrometer.core.instrument.Tag;
 import io.micrometer.core.instrument.Tags;
 import io.micrometer.core.instrument.binder.cache.HazelcastCacheMetrics;
-import java.lang.reflect.InvocationHandler;
-import java.lang.reflect.Proxy;
-import java.util.Collections;
-import java.util.List;
-import javax.annotation.Nonnull;
 import org.alfresco.enterprise.repo.cluster.cache.HazelcastSimpleCache;
 import org.alfresco.enterprise.repo.cluster.cache.InvalidatingCache;
 import org.alfresco.enterprise.repo.cluster.core.ClusterServiceInitialisedEvent;
 import org.alfresco.enterprise.repo.cluster.core.ClusteredObjectProxyFactory.ClusteredObjectProxyInvoker;
 import org.alfresco.repo.cache.DefaultSimpleCache;
 import org.alfresco.repo.cache.SimpleCache;
-import org.alfresco.service.descriptor.DescriptorService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.ApplicationEvent;
+
+import javax.annotation.Nonnull;
+import java.lang.reflect.InvocationHandler;
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
+import java.lang.reflect.Proxy;
+import java.util.Collections;
+import java.util.List;
 
 public class EnterpriseCacheMetrics implements EventTriggeredMeterBinder, NamedMeterBinder {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(EnterpriseCacheMetrics.class);
 
     private ApplicationContext ctx;
-    private DescriptorService descriptorService;
 
-    EnterpriseCacheMetrics(ApplicationContext ctx,
-            DescriptorService descriptorService) {
+
+    EnterpriseCacheMetrics(ApplicationContext ctx
+    ) {
         this.ctx = ctx;
-        this.descriptorService = descriptorService;
+
     }
 
     @Override
@@ -87,7 +89,8 @@ public class EnterpriseCacheMetrics implements EventTriggeredMeterBinder, NamedM
 
         if (!(invocationHandler instanceof ClusteredObjectProxyInvoker)) {
             LOGGER.warn(
-                    "Cache '{}': cannot resolve proxy because invocation handler of type '{}' is not an instance of 'ClusteredObjectProxyInvoker'",
+                    "Cache '{}': cannot resolve proxy because invocation handler" +
+                            " of type '{}' is not an instance of 'ClusteredObjectProxyInvoker'",
                     cacheBeanName, invocationHandler.getClass());
             return object;
         }
@@ -98,17 +101,35 @@ public class EnterpriseCacheMetrics implements EventTriggeredMeterBinder, NamedM
     private void monitorCache(final HazelcastSimpleCache cache, final MeterRegistry registry)
             throws NoSuchFieldException, IllegalAccessException {
         IMap<?, ?> cacheMap = ReflectionUtil.extractField(cache, "map");
-
-        if (getAlfrescoMajorVersion() <= 5) {
-            Hazelcast2CacheMetrics.monitor(registry, cacheMap, "type", cache.getClass().getSimpleName());
-        } else {
-            HazelcastCacheMetrics.monitor(registry, cacheMap, "type", cache.getClass().getSimpleName());
+        Method method =
+                ReflectionUtil.getMethod(
+                        HazelcastCacheMetrics.class,
+                        "monitor",
+                        MeterRegistry.class,
+                        Object.class,
+                        Iterable.class
+                );
+        if (method == null) {
+            method = ReflectionUtil.getMethod(
+                    HazelcastCacheMetrics.class,
+                    "monitor",
+                    MeterRegistry.class,
+                    IMap.class,
+                    Iterable.class
+            );
+        }
+        if (method == null) {
+            LOGGER.warn("could not acquire HazelcastCacheMetrics.monitor method, {} will not be monitored",
+                    cache.getClass().getSimpleName());
+            return;
+        }
+        try {
+            method.invoke(null, registry, cacheMap, Tags.of("type", cache.getClass().getSimpleName()));
+        } catch (InvocationTargetException e) {
+            LOGGER.error(e.getMessage(), e);
         }
     }
 
-    private int getAlfrescoMajorVersion() {
-        return Integer.valueOf(descriptorService.getServerDescriptor().getVersionMajor());
-    }
 
     private void monitorCache(final String cacheBeanName, final InvalidatingCache cache, final MeterRegistry registry)
             throws NoSuchFieldException, IllegalAccessException {