+++ alfred-telemetry-platform/src/main/java/eu/xenit/alfred/telemetry/util/ReflectionUtil.java
@@ -1,8 +1,13 @@
 package eu.xenit.alfred.telemetry.util;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import java.lang.reflect.Field;
+import java.lang.reflect.Method;
 
 public class ReflectionUtil {
+    private static final Logger LOGGER = LoggerFactory.getLogger(ReflectionUtil.class);
 
     private ReflectionUtil() {
         // private ctor to hide implicit public one
@@ -16,4 +21,12 @@ public class ReflectionUtil {
         return (T) field.get(object);
     }
 
+    public static Method getMethod(Class<?> target, String methodName, Class<?>... parameterTypes) {
+        try {
+            return target.getMethod(methodName, parameterTypes);
+        } catch (NoSuchMethodException e) {
+            LOGGER.warn(e.getMessage());
+        }
+        return null;
+    }
 }